refactor: extract GeneralAMCQuestionnaireMapper from AMCQuestionnaireD2Repository
